more changes related to changes requested in PR #204)
